Link to canonical feature request for Google Cloud SQL
Before this change, the README links to a Google Support Tracker issue that is now closed and marked as a duplicate of a different issue for requesting pgvector support. If the community would like to make it's voice heard to Google, we should pipe in on the right issue so their prioritization can see how many folks actually would like this!