Update ivfflat cost estimation to choose index for more searches
This makes several modification to the ivfflat cost estimation to
help the query planner choose it for searches in appropriate
situations.

First, the update costing algorithm estimates the number of pages
to visit by estimating the number of tuples based on the ratio
of lists to visit. Additionally, the updated costs sets the
selectivity to the probe/list ratio, given we are only visiting
the subset of tuples when the probe number is less than the number
of lists.

With these updates, we can infer the startup cost from the generic
cost estimator, as we've tuned the selectivity and page fetching
values.

Testing on data sets of different size show that the query planner
will choose the ivfflat index in additional appropriate situations
instead of defaulting back to a sequential scan.