Modify index AM definitions to follow upstream convention
The upstream index access methods declare them to be extern. While
the current definitions seem benign, this definition is needed
when an access method may be used beyond its current scope. For
example, prior to this commit, the index costing functions in
PostgreSQL would not execute the full functions due to the
ivfflatcostestimate not having an extern definition. In this case,
this would skip determining if PostgreSQL should perform a parallel
index scan.