Ivfflat index only supports scans with MVCC-compliant snapshots
Ivfflat index firstly fetch all index tuples which match the scan
keys, and sort those tuples by distance. Then, fetch the heap tuples by
tid. So the ivfflat index scans are asynchronous, which means it is only
workable for a query using an MVCC snapshot.

This pr can improve the performance as it doesn't read and lock the
buffer each time before fetching the heap tuple.  And it is not
necessary for ivfflat to support non-MVCC snapshot.