Speed up HNSW build, by memoizing some state in Algorithm 4.
We use HNSW Algorithm 4 in HNSW index build, when updating
connnections of an index. It starts with an empty set R, and greedily
adds neighbors to it, whenever the next-neareset neighbor is closer to
the element than any other neighbor already in R. That's wasteful when
we're updating the connections of an existing node. When updating, we
always add one neighbor to the candidate set, and prune the
now-weakest candidate. We do a lot of repeated work, repeatedly
building the set R with the same candidates.

To speed that up, remember which candidates are part of set R, and
which ones are kept in the neighbors list just to fill it up to size M
(see keepPrunedConnections in the paper). When we come back to rebuild
the set R, with the one extra element, we can skip doing the distance
calculations until we see the new element, because we know that they
still have the same distances as before, so they still belong to R.

This dramatically reduces the number of distance calculations
performed during index build. A downside is that we need to memorize
an extra flag for each neighbor, which increases memory usage. We only
need a bit or two, so we could easily steal those as a compact bitmap,
or steal the bits from the 'element' pointer field. Or e.g. allocate
elements in larger chunks, reducing palloc overhead, to make up for
that. But in this patch, I just added the new field in the most
straightforward manner.