Speed up HNSW build more, by biggger rewrite of update-neigbors algorithm.
This takes the approach of the previous patch further. We still use
Algorithm 4 for choosing the neighbors, but do it even more
incrementally.

WIP: I'm still seeing the "UpdateNeighbors chose different element to
prune than SelectNeighbors" WARNING very occasionally with this, when
assertions are enabled. I'm not sure if it's a bug, or just a matter
of different random choice between SelectNeighbors and UpdateNeighbors.