Provide estimate of total tuples to insert on bulk HNSW build
We won't know how many tuples will actually be inserted until we
finish the full table scan, but we can estimate the total based on
the value avaulable in `pg_class.reltuples`. This allows for the
index creator to get an approximate estimate of how long a build
may take.

Note that this leaves the default value of `-1` for reltuples in
place if there is no data available.