Refactor InsertTuple function
Rename it to InsertTupleInMemory() to make it clear that it is used
only in the in-memory phase. Move all the work to inside the function,
so that BuildCallback() is only responsible for routing to the
in-memory or on-disk insert function. Makes BuildCallback() look more
symmetric, and makes it easier to see the whole picture of the
in-memory insert case, as it's all in the same function.