Allocate HnswElement's vector value together with the struct
This reduces memory usage, by eliminating the palloc overhead. Every
palloc allocation has some constant overhead, and it also rounds up
small allocations to next power-of-two. This commit reduces the
discrepancy in the memoryLeft accounting and actual ememory usage by
having fewer small allocations.

The reason I started to hack on this originally was to keep the vector
as close as possible to 'hash' field, so that they would fit on the
same cache line as often as possible. That way in HnswSearchLayer(),
when you fetch the 'hash' to call AddToVisited(), you fetch the
beginning of the vector too. That eliminates one cache miss stall for
each visited neighbor. And I think I can see that effect in 'perf'
profile: the cache miss showed up in the pg_detoast_datum() function,
because that is the first dereference of the pointer when visiting an
element, and went down from about 15% to 2% of CPU usage. However, I'm
not seeing any meaningful overall speedup in HNSW build. So I guess
the overhead or stall moved somewhere else, but I'm not sure where. In
any case, this seems worthwhile for the reduction in memory usage
alone.