Store heap tid inline in HnswElement when possible
A List has fair amount of overhead. The List struct itself is 24
bytes, plus 8 bytes pointer and the ItemPointerData allocation plus
palloc overhead of that for each TID stored in the List.

In the common case that an HnswElement has exactly one heap TID, store
it inline in the HnswElement. If there are more than one TID, allocate
an array that can hold all HNSW_HEAPTIDS tids and use that from there
on. This only takes 8 + 1 (for array length) bytes in HnswElement when
storing one heap TID, and if there are more TIDs, 60 bytes to hold the
array. So this is much more compact in the common case of 1 TID, maybe
a little larger for 2-3 TIDs, and certainly more compact for more TIDs
again.

This brings the memory usage down, reducing the discrepancy between
memoryLeft accounting and actual memory usage.

The estimate in HnswElementMemory() used to account 6 bytes for the
single TID. In principle, that should now be removed, as a single TID
is inlined in HnswElement now. But I didn't want to change the
estimate to make it easier to see the effect of this patch, and also
because our estimate is already too small, because of other
unaccounted-for overhead. I just changed it to a constant 6 bytes
instead of sizeof(ItemPointerData), and added a comment.