Simplify relptr stuff
The HnswElementPtr type was weird. It was a union of the relptr type
and the non-relative pointer type. Seems simpler to always use
relptrs.

From a performance point of view it's a wash. In HnswPtrAccess() this
saves 1-2 instructions, from not having to check for base == NULL, but
introduces one instruction to add 1 to the offset to get the pointer.