Remove MAXALIGN from SparseVector
A SparseVector as whole is only 4-byte aligned, so it's a bit weird to
add padding to 8 bytes (on 64-bit platforms) in the middle of the
struct. It doesn't guarantee that that the values would land on an
8-byte boundary anyway.

For clarity, also turn SPARSEVEC_SIZE and SPARSEVEC_VALUES into static
inline functions. It's more tidy to add comments on proper functions.
Also eliminates the double-evaluation hazard of the _nnz argument.