Using NEON (if available) for calculating l2_distance and inner_product
This commit adds support for the NEON SIMB extension. This extension is
a bit older but it is supported by all current AWS Gravaton chips.
Gravaton 3 and Gravaton 4 support SVE and SVE 2 respectively but this
commit should still offer a sizable performance improvement across the
board.

Early testing (on a Macbook Pro) has shown a roughly 8x performance
improvement at the cost of roughly 0.0005% precision. This is a result
of the order in how floating point values are calculated and it seemed
like a reasonable tradeoff for the increased performance.